aya: switch to rustified enums

+++ aya/src/bpf.rs
@@ -80,7 +80,8 @@ impl Bpf {
 
         let mut maps = Vec::new();
         for (_, mut obj) in obj.maps.drain() {
-            if obj.def.map_type == BPF_MAP_TYPE_PERF_EVENT_ARRAY && obj.def.max_entries == 0 {
+            if obj.def.map_type == BPF_MAP_TYPE_PERF_EVENT_ARRAY as u32 && obj.def.max_entries == 0
+            {
                 obj.def.max_entries = *possible_cpus()
                     .map_err(|error| BpfError::FileError {
                         path: PathBuf::from(POSSIBLE_CPUS),
