+++ aya/src/maps/program_array.rs
@@ -22,7 +22,7 @@ pub struct ProgramArray<T: Deref<Target = Map>> {
 impl<T: Deref<Target = Map>> ProgramArray<T> {
     pub fn new(map: T) -> Result<ProgramArray<T>, MapError> {
         let map_type = map.obj.def.map_type;
-        if map_type != BPF_MAP_TYPE_PROG_ARRAY {
+        if map_type != BPF_MAP_TYPE_PROG_ARRAY as u32 {
             return Err(MapError::InvalidMapType {
                 map_type: map_type as u32,
             })?;
